@@ -15,8 +15,8 @@ class ShouldBeVerified
      */
     public function handle(Request $request, Closure $next): Response
     {
-        if ($request->user()->email_verified_at == null) {
-            return to_route('auth.email-validation');
+        if ($request->user()->email_verified_at === null) {
+            return to_route('email-validation');
         }
 
         return $next($request);
