@@ -6,28 +6,28 @@ use App\Events\SendNewCode;
 use App\Models\User;
 use App\Notifications\WelcomeNotification;
 use App\Rules\CodeValidation;
+use App\Traits\Livewire\HasToast;
 use Carbon\Carbon;
 use Illuminate\View\View;
-use Livewire\Attributes\{Layout, On, Rule};
+use Livewire\Attributes\{Layout, On, Validate};
 use Livewire\Component;
 
 class EmailValidation extends Component
 {
-    #[Rule(new CodeValidation())]
+    use HasToast;
+
+    #[Validate(['required', new CodeValidation])]
     public ?string $code = null;
 
     public ?Carbon $dateResendAllowed = null;
 
-    #[Layout('components.layouts.guest')]
-    public function render(): View
-    {
-        return view('livewire.auth.email-validation')->title('Verify authentication code');
-    }
+    public bool $welcome = false;
 
     public function handle(): void
     {
+        $this->validate();
+
         try {
-            $this->validate();
 
             /* @var User $user */
             $user = auth()->user();
@@ -36,30 +36,50 @@ class EmailValidation extends Component
             $user->validation_code   = null;
             $user->save();
 
-            $user->notify(new WelcomeNotification());
+            if ($this->welcome) {
+                $user->notify(new WelcomeNotification);
+            }
 
-            $this->redirect(route('dashboard'));
+            $this->redirect(route('home'));
         } catch (\Exception $e) {
-            $this->addError('code', $e->getMessage());
+            $error_message = __('messages.invalid_code');
+            $this->addError('invalidCode', $error_message);
+            $this->warning($error_message);
         }
     }
 
     public function sendNewCode(): void
     {
         if ($this->dateResendAllowed && now()->isBefore($this->dateResendAllowed)) {
-            $this->addError('dateResendNotAllowed', 'Wait 3 minutes for another attempt');
+            $error_message = __('messages.resend_code_attempt', ['number' => '3']);
+            $this->addError('dateResendNotAllowed', $error_message);
+            $this->warning($error_message);
 
             return;
         }
 
         SendNewCode::dispatch(auth()->user());
+        $this->success(__('messages.code_sent_successfully'));
 
         $this->dateResendAllowed = now()->addMinutes(3);
     }
 
     #[On('resend-allowed')]
-    public function resendAllowed()
+    public function resendAllowed(): void
     {
         $this->dateResendAllowed = null;
+    }
+
+    public function mount(): void
+    {
+        if (isset(request()->welcome)) {
+            $this->welcome = true;
+        }
+    }
+
+    #[Layout('components.layouts.guest')]
+    public function render(): View
+    {
+        return view('livewire.auth.email-validation')->title('Verify authentication code');
     }
 }
