@@ -2,32 +2,36 @@
 
 namespace App\Livewire\Auth\Password;
 
+use Flux\Flux;
 use Illuminate\Support\Facades\Password;
 use Illuminate\View\View;
-use Livewire\Attributes\Rule;
+use Livewire\Attributes\Validate;
 use Livewire\Component;
 
 class Recovery extends Component
 {
-    #[Rule(['required', 'email'])]
+    #[Validate(['required', 'email'])]
     public ?string $email = null;
 
-    public ?string $message = null;
+    public function submit(): void
+    {
+        $this->validate();
+
+        Password::sendResetLink($this->only('email'));
+
+        Flux::toast(
+            text: __('messages.success'),
+            heading: __('messages.password_recovery_sent_link'),
+            variant: 'success'
+        );
+
+        $this->reset('email');
+    }
 
     public function render(): view
     {
         return view('livewire.auth.password.recovery')
             ->title('Password recovery')
             ->layout('components.layouts.guest');
-    }
-
-    public function recoveryPassword()
-    {
-        $this->validate();
-
-        Password::sendResetLink($this->only('email'));
-
-        $this->message = "You will receive an email with the password recovery link.";
-        $this->reset('email');
     }
 }
