@@ -3,21 +3,27 @@
 namespace App\Models;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
+use App\Enums\Status;
+use App\Traits\Models\BelongsToTenant;
 use App\Traits\Models\HasPermissions;
+use App\Traits\Models\HasSearch;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
+use Illuminate\Support\Str;
 use OwenIt\Auditing\Contracts\Auditable;
 
 class User extends Authenticatable implements Auditable
 {
     use HasFactory;
-    use Notifiable;
     use HasPermissions;
-    use SoftDeletes;
+    use Notifiable;
     use \OwenIt\Auditing\Auditable;
+    use SoftDeletes;
+    use HasSearch;
+//    use BelongsToTenant;
 
     /**
      * The attributes that are mass assignable.
@@ -32,6 +38,8 @@ class User extends Authenticatable implements Auditable
         'deleted_by',
         'restored_at',
         'restored_by',
+        'level',
+        'status',
     ];
 
     /**
@@ -54,7 +62,20 @@ class User extends Authenticatable implements Auditable
         return [
             'email_verified_at' => 'datetime',
             'password'          => 'hashed',
+            'status' => Status::class,
         ];
+    }
+
+    /**
+     * Get the user's initials
+     */
+    public function initials(): string
+    {
+        return Str::of($this->name)
+            ->explode(' ')
+            ->take(2)
+            ->map(fn ($word) => Str::substr($word, 0, 1))
+            ->implode('');
     }
 
     public function deletedBy(): BelongsTo
