@@ -2,9 +2,6 @@
 
 namespace App\Providers;
 
-use App\Enums\Can;
-use App\Models\User;
-use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
@@ -22,11 +19,6 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        foreach (Can::cases() as $can) {
-            Gate::define(
-                $can->value,
-                fn (User $user) => $user->hasPermissionTo($can)
-            );
-        }
+        //
     }
 }
