@@ -1,18 +1,88 @@
 <?php
 
 return [
+    'actions' => 'Actions',
+    'active' => 'Active',
     'already_have_an_account' => 'Already have an account?',
+    'appearance' => 'Appearance',
+    'back' => 'Back',
+    'back_to' => 'Back to',
+    'blocked' => 'Blocked',
+    'change_status' => 'Change status',
+    'code' => 'Code',
+    'code_sent_successfully' => 'Code sent successfully',
+    'confirm_email' => 'Confirm email',
     'confirm_password' => 'Confirm password',
     'continue_with_github' => 'Continue with github',
     'continue_with_google' => 'Continue with google',
+    'create' => 'Create',
+    'create_account' => 'Create account',
+    'current_password' => 'Current password',
+    'danger' => 'Danger!',
+    'delete' => 'Delete',
+    'edit' => 'Edit',
     'email' => 'Email',
+    'email_changes_must_be_verified_again' => 'Email changes must be verified again',
+    'email_verification_instruction' => 'We have sent a verification code to your email. Please enter the token below to proceed',
+    'english' => 'English',
+    'ensure_your_account_is_using_a_long_random_password_to_stay_secure' => 'Ensure your account is using a long random password to stay secure',
+    'enter_your_email' => 'Enter your email',
+    'enter_your_password' => 'Enter your password',
+    'filters' => 'Filters',
     'first_time_around_here' => 'First time around here?',
     'forgot_password' => 'Forgot password?',
+    'home' => 'Home',
+    'identification_number' => 'Identification number',
+    'inactive' => 'Inactive',
+    'invalid_code' => 'Invalid code',
+    'invalid_current_password' => 'Invalid current password',
+    'invalid_locale' => 'Invalid locale',
+    'invalid_password_reset_link' => 'The password reset link is invalid or has expired. Please request a new one',
+    'keep_connect' => 'To keep connected with us please login with your personal info',
+    'language' => 'Language',
     'login' => 'Login',
+    'login_into_your_account' => 'Login into your account',
+    'logout' => 'Logout',
+    'manage_your_profile_and_account_settings' => 'Manage your profile and account settings',
     'name' => 'Name',
+    'new_tenant' => 'New tenant',
     'password' => 'Password',
+    'password_recovery' => 'Password recovery',
+    'password_recovery_instruction' => 'Provide the email address associated with your account to continue',
+    'password_recovery_sent_link' => 'You will receive an email with the password recovery link',
+    'password_updated_successfully' => 'Password updated successfully',
+    'pending' => 'Pending',
+    'portuguese' => 'Portuguese',
+    'preferences' => 'Preferences',
+    'profile' => 'Profile',
+    'profile_updated_successfully' => 'Profile updated successfully',
+    'recover_password' => 'Recover password',
     'register' => 'Register',
     'remember_me' => 'Remember me',
-    'signup_for_free' => 'Sign up for free',
-    'warning' => 'Warning',
+    'resend_code_attempt' => 'Wait :number minutes for another attempt',
+    'resend_code_in' => 'Resend code in',
+    'reset' => 'Reset',
+    'reset_password' => 'Reset password',
+    'save' => 'Save',
+    'search' => 'Search',
+    'select_your_preferred_language_for_the_interface' => 'Select your preferred language for the interface',
+    'send_email' => 'Send email',
+    'send_new_code' => 'Send a new code',
+    'settings' => 'Settings',
+    'show' => 'Show',
+    'sign_in' => 'Sign in',
+    'sign_up' => 'Sign up',
+    'social_reason' => 'Social reason',
+    'status' => 'Status',
+    'subdomain' => 'Subdomain',
+    'success' => 'Success!',
+    'tenants' => 'Tenants',
+    'update_the_appearance_settings_for_your_account' => 'Update the appearance settings for your account',
+    'update_your_name_and_email_address' => 'Update your name and email address',
+    'verify' => 'Verify',
+    'verify_your_email' => 'Verify your email',
+    'warning' => 'Warning!',
+    'welcome_back' => 'Welcome back',
+    'welcome_home' => 'Welcome to :project',
+    'welcome_user' => 'Hi, :name. Welcome back',
 ];
