@@ -1,18 +1,88 @@
 <?php
 
 return [
+    'actions' => 'Ações',
+    'active' => 'Ativo',
     'already_have_an_account' => 'Já possui uma conta?',
+    'appearance' => 'Aparência',
+    'back' => 'Voltar',
+    'back_to' => 'Voltar para',
+    'blocked' => 'Bloqueado',
+    'change_status' => 'Alterar status',
+    'code' => 'Código',
+    'code_sent_successfully' => 'Código enviado com sucesso',
+    'confirm_email' => 'Confirmar e-mail',
     'confirm_password' => 'Confirmar senha',
     'continue_with_github' => 'Continue com o GitHub',
     'continue_with_google' => 'Continue com o Google',
+    'create' => 'Cadastrar',
+    'create_account' => 'Criar uma conta',
+    'current_password' => 'Senha atual',
+    'danger' => 'Cuidado!',
+    'delete' => 'Excluir',
+    'edit' => 'Editar',
     'email' => 'E-mail',
+    'email_changes_must_be_verified_again' => 'Se o e-mail for alterado, será necessário verificá-lo novamente',
+    'email_verification_instruction' => 'Enviamos um código de verificação para o seu e-mail. Por favor, insira o token abaixo para continuar',
+    'english' => 'Inglês',
+    'ensure_your_account_is_using_a_long_random_password_to_stay_secure' => 'Verifique se sua conta está usando uma senha aleatória longa para se manter seguro',
+    'enter_your_email' => 'Digite seu e-mail',
+    'enter_your_password' => 'Digite sua senha',
+    'filters' => 'Filtros',
     'first_time_around_here' => 'Primeira vez por aqui?',
     'forgot_password' => 'Esqueceu sua senha?',
+    'home' => 'Inicio',
+    'identification_number' => 'CNPJ',
+    'inactive' => 'Inativo',
+    'invalid_code' => 'Código inválido',
+    'invalid_current_password' => 'Senha atual inválida',
+    'invalid_locale' => 'Idioma inválido',
+    'invalid_password_reset_link' => 'O link de redefinição de senha é inválido ou expirou. Por favor, solicite um novo',
+    'keep_connect' => 'Para continuar conectado conosco, faça login com suas informações pessoais',
+    'language' => 'Idioma',
     'login' => 'Entrar',
+    'login_into_your_account' => 'Faça login em sua conta',
+    'logout' => 'Sair',
+    'manage_your_profile_and_account_settings' => 'Gerencie seu perfil e configurações de conta',
     'name' => 'Nome',
+    'new_tenant' => 'Novo inquilino',
     'password' => 'Senha',
+    'password_recovery' => 'Recuperar senha',
+    'password_recovery_instruction' => 'Informe o endereço de e-mail associado à sua conta para continuar',
+    'password_recovery_sent_link' => 'Você receberá um e-mail com o link de recuperação de senha',
+    'password_updated_successfully' => 'Senha atualizada com sucesso',
+    'pending' => 'Pendente',
+    'portuguese' => 'Português',
+    'preferences' => 'Preferências',
+    'profile' => 'Perfil',
+    'profile_updated_successfully' => 'Perfil atualizado com sucesso',
+    'recover_password' => 'Recuperar senha',
     'register' => 'Registrar',
     'remember_me' => 'Lembrar me',
-    'signup_for_free' => 'Inscreva se gratuitamente',
-    'warning' => 'Aviso',
+    'resend_code_attempt' => 'Aguarde :number minutos para outra tentativa',
+    'resend_code_in' => 'Reenviar código em',
+    'reset' => 'Redefinir',
+    'reset_password' => 'Redefinir senha',
+    'save' => 'Salvar',
+    'search' => 'Pesquisar',
+    'select_your_preferred_language_for_the_interface' => 'Selecione seu idioma preferido para sua interface',
+    'send_email' => 'Enviar e-mail',
+    'send_new_code' => 'Enviar novo código',
+    'settings' => 'Configurações',
+    'show' => 'Visualizar',
+    'sign_in' => 'Entrar',
+    'sign_up' => 'Inscreva se',
+    'social_reason' => 'Razão social',
+    'status' => 'Status',
+    'subdomain' => 'Subdomínio',
+    'success' => 'Sucesso!',
+    'tenants' => 'Inquilinos',
+    'update_the_appearance_settings_for_your_account' => 'Atualize as configurações de aparência para sua conta',
+    'update_your_name_and_email_address' => 'Atualize seu nome e endereço de e -mail',
+    'verify' => 'Verificar',
+    'verify_your_email' => 'Verifique seu e-mail',
+    'warning' => 'Aviso!',
+    'welcome_back' => 'Bem vindo de volta',
+    'welcome_home' => 'Bem vindo ao :project',
+    'welcome_user' => 'Oi, :name. Bem-vindo de volta',
 ];
