@@ -1,43 +1,31 @@
-{{--<x-card title="Code" class="w-[450px]">--}}
+<div class="py-24">
 
-{{--    @error('dateResendNotAllowed')--}}
-{{--    <x-alert icon="o-exclamation-triangle" class="alert-error my-4">--}}
-{{--        {{ $message }}--}}
-{{--    </x-alert>--}}
-{{--    @enderror--}}
+    <div class="text-center mb-10">
+        <flux:heading size="xl" class="text-3xl font-extrabold" accent>
+            {{ __('messages.verify_your_email') }}
+        </flux:heading>
+        <flux:subheading>{{ __('messages.email_verification_instruction') }}</flux:subheading>
+    </div>
 
-{{--    <x-form class="space-y-6 px-2" wire:submit="handle" no-separator>--}}
-{{--        <div class="flex flex-col">--}}
-{{--            <x-pin wire:model="code" size="6" numeric/>--}}
-{{--            @error('code')--}}
-{{--            <p class="text-xs text-red-500">--}}
-{{--                {{ $message }}--}}
-{{--            </p>--}}
-{{--            @enderror--}}
-{{--        </div>--}}
-{{--        <x-slot:actions>--}}
-{{--            <x-button label="Verify" class="btn-primary w-full" type="submit" spinner="handle"/>--}}
-{{--        </x-slot:actions>--}}
-{{--    </x-form>--}}
+    <form wire:submit="handle" class="space-y-4">
+        @csrf
 
-{{--    @if(!$dateResendAllowed || now()->isAfter($dateResendAllowed))--}}
-{{--        <p wire:click="sendNewCode"--}}
-{{--           class="mt-5 text-center text-sm font-semibold--}}
-{{--       {{ $dateResendAllowed && now()->isBefore($dateResendAllowed) ? 'text-gray-400' : 'cursor-pointer text-indigo-400 hover:text-indigo-300' }} "--}}
-{{--        >--}}
-{{--            Send a new code--}}
-{{--        </p>--}}
-{{--    @else--}}
-{{--        <p class="flex mt-5 justify-center text-sm font-semibold">--}}
-{{--            <span class="text-gray-400">--}}
-{{--                Resend code in--}}
-{{--                <x-timer :minutes="3"/>--}}
-{{--            </span>--}}
-{{--        </p>--}}
-{{--    @endif--}}
+        <flux:field>
+            <flux:label>{{ __('messages.code') }}</flux:label>
+            <flux:input wire:model="code"
+                        placeholder="{{ __('messages.code') }}"/>
+            <flux:error name="code"/>
+        </flux:field>
 
-{{--    <p @click="$dispatch('logout')"--}}
-{{--       class="mt-10 text-center text-sm text-gray-400 hover:text-indigo-300 cursor-pointer">--}}
-{{--        Logout--}}
-{{--    </p>--}}
-{{--</x-card>--}}
+        <flux:button type="submit" variant="primary"
+                     class="w-full uppercase cursor-pointer">{{ __('messages.verify') }}</flux:button>
+    </form>
+
+    <div class="text-center mt-10">
+       <x-timer minutes="3" />
+
+        <flux:subheading class="text-center mt-5 cursor-pointer hover:text-accent">
+            <p @click="$dispatch('logout')">{{ __('messages.logout') }}</p>
+        </flux:subheading>
+    </div>
+</div>
